extract navigable group item and nested description
