+++ client/src/components/TestSuite/TestSuiteDetails/TestGroupListItem/NestedDescriptionPanel.tsx
@@ -1,0 +1,65 @@
+import React, { FC } from 'react';
+import useStyles from './styles';
+import {
+  Accordion,
+  AccordionDetails,
+  AccordionSummary,
+  Box,
+  Divider,
+  List,
+  ListItem,
+  ListItemText,
+  Typography,
+} from '@mui/material';
+import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
+import { TestGroup } from '~/models/testSuiteModels';
+import ReactMarkdown from 'react-markdown';
+
+interface NestedDescriptionPanelProps {
+  testGroup: TestGroup;
+}
+
+const NestedDescriptionPanel: FC<NestedDescriptionPanelProps> = ({ testGroup }) => {
+  const styles = useStyles();
+
+  return (
+    <Box className={styles.nestedDescriptionContainer}>
+      <Accordion
+        disableGutters
+        key={`${testGroup.id}-description`}
+        className={styles.accordion}
+        TransitionProps={{ unmountOnExit: true }}
+      >
+        <AccordionSummary
+          id={`${testGroup.id}-description-summary`}
+          aria-controls={`${testGroup.id}-description-detail`}
+          expandIcon={<ExpandMoreIcon sx={{ padding: '0 5px' }} />}
+          sx={{ userSelect: 'auto' }}
+        >
+          <List sx={{ p: 0 }}>
+            <ListItem sx={{ p: 0 }}>
+              <ListItemText
+                primary={
+                  <Typography className={styles.nestedDescriptionHeader}>
+                    About {testGroup.short_title || testGroup.title}
+                  </Typography>
+                }
+              />
+            </ListItem>
+          </List>
+        </AccordionSummary>
+        <Divider />
+        <AccordionDetails
+          title={`${testGroup.id}-description-detail`}
+          className={styles.accordionDetailContainer}
+        >
+          <ReactMarkdown className={`${styles.accordionDetail} ${styles.nestedDescription}`}>
+            {testGroup.description as string}
+          </ReactMarkdown>
+        </AccordionDetails>
+      </Accordion>
+    </Box>
+  );
+};
+
+export default NestedDescriptionPanel;
