+++ client/src/components/TestSuite/TestSuiteDetails/TestGroupListItem/styles.tsx
@@ -1,0 +1,67 @@
+import { Theme } from '@mui/material/styles';
+import makeStyles from '@mui/styles/makeStyles';
+
+export default makeStyles((theme: Theme) => ({
+  root: {
+    backgroundColor: theme.palette.background.paper,
+  },
+  labelText: {
+    display: 'inline',
+  },
+  shortId: {
+    display: 'inline',
+    fontWeight: 'bold',
+    alignSelf: 'center',
+    color: theme.palette.common.grayDarkest,
+  },
+  currentItem: {
+    fontWeight: 600,
+  },
+  accordion: {
+    '&:before': {
+      display: 'none',
+    },
+    '&:not(last-child)': {
+      borderBottom: `1px solid ${theme.palette.divider}`,
+    },
+    '@media print': {
+      borderBottom: 'none !important',
+    },
+  },
+  accordionSummary: {
+    userSelect: 'auto',
+    '@media print': {
+      minHeight: 'unset',
+    },
+    '& .MuiAccordionSummary-content': {
+      margin: 0,
+    },
+  },
+  accordionDetailContainer: {
+    backgroundColor: theme.palette.common.grayLighter,
+    '@media print': {
+      padding: '0 0 0 16px',
+    },
+  },
+  accordionDetail: {
+    backgroundColor: theme.palette.common.white,
+    padding: 0,
+    margin: '8px 0 0 0',
+    borderRadius: '4px',
+    '@media print': {
+      margin: 0,
+    },
+  },
+  nestedDescription: {
+    padding: '8px 16px 24px 16px',
+    overflow: 'auto',
+  },
+  nestedDescriptionHeader: {
+    fontWeight: 'bolder !important',
+    color: theme.palette.common.orangeDarker,
+  },
+  nestedDescriptionContainer: {
+    backgroundColor: theme.palette.common.grayLighter,
+    padding: '8px 0',
+  },
+}));
